AlertDeleting added

+++ lib/presentation/screens/post_list_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:flutter_home_work11/domain/store/post_store.dart';
+import 'package:flutter_home_work11/presentation/widgets/custom_app_bar.dart';
 
 class PostListScreen extends StatelessWidget {
   final PostStore postStore = PostStore();
@@ -10,17 +11,15 @@ class PostListScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text("Posts")),
+      appBar: const CustomAppBar(title: 'Posts'),
       body: Observer(
         builder: (_) {
           if (postStore.isLoading) {
             return const Center(child: CircularProgressIndicator());
           }
-
           if (postStore.errorMessage != null) {
-            return Center(child: Text("Error: ${postStore.errorMessage}"));
+            return Center(child: Text('Error: ${postStore.errorMessage}'));
           }
-
           return ListView.builder(
             itemCount: postStore.posts.length,
             itemBuilder: (context, index) {
@@ -32,20 +31,49 @@ class PostListScreen extends StatelessWidget {
                   title: Text('Name: ${post.title}'),
                   subtitle: Text('Post: ${post.body}'),
                   trailing: IconButton(
-                    icon: const Icon(Icons.delete_forever,
-                        color: Colors.deepOrange,),
+                    icon: const Icon(
+                      Icons.delete_forever,
+                      color: Colors.deepOrange,
+                    ),
                     onPressed: () async {
                       final scaffoldMessenger = ScaffoldMessenger.of(context);
-                      await postStore.deletePost(post.id);
-                      scaffoldMessenger.showSnackBar(
-                        SnackBar(
-                          content: Text(
-                            postStore.errorMessage != null
-                                ? "Failed to delete post"
-                                : "Post deleted successfully!",
-                          ),
-                        ),
+                      final shouldDelete = await showDialog<bool>(
+                        context: context,
+                        builder: (BuildContext context) {
+                          return AlertDialog(
+                            title: const Text('Confirm Deletion'),
+                            content: const Text(
+                              'Are you sure you want to delete this post?',
+                            ),
+                            actions: [
+                              TextButton(
+                                onPressed: () {
+                                  Navigator.of(context).pop(false);
+                                },
+                                child: const Text('Cancel'),
+                              ),
+                              TextButton(
+                                onPressed: () {
+                                  Navigator.of(context).pop(true);
+                                },
+                                child: const Text('Delete'),
+                              ),
+                            ],
+                          );
+                        },
                       );
+                      if (shouldDelete == true) {
+                        await postStore.deletePost(post.id);
+                        scaffoldMessenger.showSnackBar(
+                          SnackBar(
+                            content: Text(
+                              postStore.errorMessage != null
+                                  ? 'Failed to delete post'
+                                  : 'Post deleted successfully!',
+                            ),
+                          ),
+                        );
+                      }
                     },
                   ),
                   onTap: () {
